feat: implement API request handling with file parsing options and validation
